+++ src/test/java/BaseTest.java
@@ -4,6 +4,8 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
+import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.BeforeSuite;
@@ -15,6 +17,7 @@ public class BaseTest {
 
             public WebDriver driver = null;
             public String url = "https://qa.koel.app/";
+            public WebDriverWait wait;
 
             @BeforeSuite
             static void setupClass() {
@@ -28,7 +31,7 @@ public class BaseTest {
                 options.addArguments("--remote-allow-origins=*");
 
                 driver = new ChromeDriver(options);
-                driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+                wait = new WebDriverWait(driver, Duration.ofSeconds(10));
                 driver.manage().window().maximize();
                 url=BaseURL;
                 navigateToPage();
@@ -44,21 +47,20 @@ public class BaseTest {
             }
 
             public void provideEmail(String email) {
-                WebElement emailField = driver.findElement(By.cssSelector("input[type='email']"));
+                WebElement emailField = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("input[type='email']")));
                 emailField.clear();
                 emailField.sendKeys(email);
 
             }
 
             public void providePassword(String password) {
-
-                WebElement passwordField = driver.findElement(By.cssSelector("input[type='password']"));
+                WebElement passwordField = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("input[type='password']")));
                 passwordField.clear();
                 passwordField.sendKeys(password);
             }
 
             public void clickSubmit() {
-                WebElement submit = driver.findElement(By.cssSelector("button[type='submit']"));
+                WebElement submit = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("button[type='submit']")));
                 submit.click();
             }
         }
