+++ src/test/resources/TestNG.xml
@@ -5,7 +5,7 @@
 <parameter name="BaseURL" value="https://qa.koel.app/"/>
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="Homework19Test"/>
+            <class name="Homework20Test"/>
         </classes>
     </test>
 </suite>
